Added device list page

+++ app/src/main/java/com/example/irrigation/DeviceData.kt
@@ -2,7 +2,7 @@ package com.example.irrigation
 
 data class DeviceData(
     val number_of_devices: Int = 0,
-    val device_list: List<Int> = listOf(),
+    val device_list: List<Byte> = listOf(),
     val motor_state: String = "OFF",
     val rem_time: Int = 0,
     val valve0_state: String = "Closed",
